Move dataverse integration code to lib.dataverse
